feat(setup): ✨ Add development setup documentation and improve import structure
- Created `DEVELOPMENT.md` for clear project setup instructions.
- Removed `sys.path.append()` calls across multiple scripts for cleaner imports.
- Updated `pyproject.toml` to configure pytest paths and options.
- Adjusted `requirements.txt` to specify compatible version for `ocha-stratus`.
- Enhanced import statements in various modules to use absolute imports from the `src` package.
