+++ requirements.txt
@@ -4,4 +4,4 @@ beautifulsoup4>=4.12.0
 pandas>=2.0.0
 PyPDF2>=3.0.0
 openai>=1.0.0
-ocha-stratus>=1.0.0
+ocha-stratus>=0.1.0
